add swagger DOC for Admin/Market/CommonDiscountController

+++ app/Http/Requests/ApiRequests/Admin/Market/CommonDiscount/CommonDiscountStoreApiRequest.php
@@ -23,11 +23,11 @@ class CommonDiscountStoreApiRequest extends ApiFormRequest
     {
         return [
             'title' => ['required','min:2','max:120'],
-            'percentage' => ['required','numeric','min:1','max:100'],
-            'discount_ceiling' => ['nullable','numeric'],
-            'minimum_order_amount' => ['nullable','numeric'],
-            'start_date' => ['required','numeric'],
-            'end_date' => ['required','numeric'],
+            'percentage' => ['required','integer','min:1','max:100'],
+            'discount_ceiling' => ['nullable','integer'],
+            'minimum_order_amount' => ['nullable','integer'],
+            'start_date' => ['required','integer'],
+            'end_date' => ['required','integer'],
         ];
     }
 }
